backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
supabase==2.3.0
face-recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
librosa==0.10.1
pyaudio==0.2.14
openai-whisper==20231117
scikit-learn==1.3.2
joblib==1.3.2
transformers==4.36.2
torch==2.1.2
spacy==3.7.2
pyttsx3==2.90
